Update README.md
This project involves web scraping key events of the 20th century from a Wikipedia page and analyzing the frequency with which various countries are mentioned in these events. The analysis includes generating a plot that visualizes how many times each country is mentioned, specifically focusing on countries that are mentioned more than once.